Move `CoreDataNewsStore` implementation of the `NewsStore` protocol to it's own file
