+++ CleanNewsFramework/News Cache/CoreDataNewsStore.swift
@@ -7,50 +7,15 @@
 
 import CoreData
 
-public final class CoreDataNewsStore: NewsStore {
+public final class CoreDataNewsStore {
     private let container: NSPersistentContainer
     private let context: NSManagedObjectContext
-    
+
     public init(storeURL: URL, bundle: Bundle = .main) throws {
         container = try NSPersistentContainer.load(modelName: "NewsStore", url: storeURL, in: bundle)
         context = container.newBackgroundContext()
     }
     
-    public func retrieve(completion: @escaping RetrievalCompletion) {
-        let context = self.context
-        context.perform {
-            completion(Result {
-                try ManagedCache.find(in: context).map {
-                    CachedNews(items: $0.localNews, timestamp: $0.timestamp)
-                }
-            })
-        }
-    }
-    
-    public func insert(_ items: [LocalNewsItem], timestamp: Date, completion: @escaping InsertionCompletion) {
-        let context = self.context
-        
-        context.perform {
-            completion(Result {
-                let managedCache = try ManagedCache.newUniqueInstance(in: context)
-                managedCache.timestamp = timestamp
-                managedCache.newsFeed = ManagedNewsItem.articles(from: items, in: context)
-                
-                try context.save()
-            })
-        }
-    }
-    
-    public func deleteCachedNews(completion: @escaping DeletionCompletion) {
-        let context = self.context
-        
-        context.perform {
-            completion(Result {
-                try ManagedCache.find(in: context).map(context.delete).map(context.save)
-            })
-        }
-    }
-
     func perform(_ action: @escaping (NSManagedObjectContext) -> Void) {
         let context = self.context
         context.perform {
